created modal
created a add timed message modal to make the timed messages font end look better
